Combine Huffman tests
